fixed one line json script to look for valid rule property for katago ai, moved around some of the learn and practice pages and added the correct db schema
