+++ frontend/src/components/LearnAndPractice/BasicRulesPage3.js
@@ -1,15 +1,25 @@
 import React from "react";
 
 const BasicRulesPage3 = () => {
-
   return (
     <div className="tutorial-puzzle-page">
       <h2>Basic Rules: Placing Stones & Surrounding Territory 3</h2>
-      <p>Who won the game below and by how much?</p>
-      <p>*** For later: you may be wondering why black or white can't just invade
-        the opponent's territory.  That's an excellent thought, and we will explore that
-        later in the game end condition and two eyes section
+      <p>Who won the game below and by how much? 🤔</p>
+      <p>
+        If you are stuck, please refer to the previous pages. When you are
+        ready, please go to the next page where we discuss the answer!
       </p>
+      <p>
+        You may be wondering why black or white can't just invade the opponent's
+        territory. We will explore that later, but for now let's focus on
+        learning from these simpler examples.
+      </p>
+      {/* <div className="image-text-wrap"> */}
+      <img
+        src="/learn-and-practice-images/basic-rules-5.png"
+        alt="basic-rules-5"
+      />
+      {/* </div> */}
     </div>
   );
 };
