+++ frontend/src/components/LearnAndPractice/LearnAndPracticeBasicRules.js
@@ -7,6 +7,7 @@ import BasicRulesPage2 from "./BasicRulesPage2";
 import BasicRulesPage3 from "./BasicRulesPage3";
 import BasicRulesPage4 from "./BasicRulesPage4";
 import BasicRulesPage5 from "./BasicRulesPage5";
+import BasicRulesPage6 from "./BasicRulesPage6";
 
 const LearnAndPracticeBasicRules = () => {
   const history = useHistory();
@@ -44,6 +45,7 @@ const LearnAndPracticeBasicRules = () => {
         pageNumber === '3' ? <BasicRulesPage3 /> :
         pageNumber === '4' ? <BasicRulesPage4 /> :
         pageNumber === '5' ? <BasicRulesPage5 /> :
+        pageNumber === '6' ? <BasicRulesPage6 /> :
         <div>Page not found</div>
       }
       <button onClick={navigateToNextPage} className="next-page-button">
